chore: move/remove methods from TemplateManager
not necessary anymore
